+++ src/Relationships/HasOneToMany.php
@@ -2,7 +2,7 @@
 
 namespace Somnambulist\Components\ReadModels\Relationships;
 
-use Somnambulist\Collection\MutableCollection as Collection;
+use Somnambulist\Collection\Contracts\Collection;
 use Somnambulist\Components\ReadModels\Manager;
 use Somnambulist\Components\ReadModels\Model;
 use Somnambulist\Components\ReadModels\ModelBuilder;
@@ -64,7 +64,7 @@ class HasOneToMany extends HasOneOrMany
                 }
             }
 
-            $model->setRelationshipValue($relationship, new Collection($entities));
+            $model->setRelationshipValue($relationship, $this->related->getCollection($entities));
         });
 
         return $this;